+++ pkg/querier/querier.go
@@ -71,6 +71,7 @@ type Config struct {
 var (
 	errBadLookbackConfigs                             = errors.New("bad settings, query_store_after >= query_ingesters_within which can result in queries not being sent")
 	errShuffleShardingLookbackLessThanQueryStoreAfter = errors.New("the shuffle-sharding lookback period should be greater or equal than the configured 'query store after'")
+	errEmptyTimeRange                                 = errors.New("empty time range")
 )
 
 // RegisterFlags adds the flags required to config this to the given FlagSet.
@@ -204,24 +205,26 @@ func NewQueryable(distributor QueryableWithFilter, stores []QueryableWithFilter,
 	return storage.QueryableFunc(func(ctx context.Context, mint, maxt int64) (storage.Querier, error) {
 		now := time.Now()
 
-		if cfg.MaxQueryIntoFuture > 0 {
-			maxQueryTime := util.TimeToMillis(now.Add(cfg.MaxQueryIntoFuture))
+		userID, err := user.ExtractOrgID(ctx)
+		if err != nil {
+			return nil, err
+		}
 
-			if mint > maxQueryTime {
-				return storage.NoopQuerier(), nil
-			}
-			if maxt > maxQueryTime {
-				maxt = maxQueryTime
-			}
+		mint, maxt, err = validateQueryTimeRange(ctx, userID, mint, maxt, limits, cfg.MaxQueryIntoFuture)
+		if err == errEmptyTimeRange {
+			return storage.NoopQuerier(), nil
+		} else if err != nil {
+			return nil, err
 		}
 
 		q := querier{
-			ctx:              ctx,
-			mint:             mint,
-			maxt:             maxt,
-			chunkIterFn:      chunkIterFn,
-			tombstonesLoader: tombstonesLoader,
-			limits:           limits,
+			ctx:                ctx,
+			mint:               mint,
+			maxt:               maxt,
+			chunkIterFn:        chunkIterFn,
+			tombstonesLoader:   tombstonesLoader,
+			limits:             limits,
+			maxQueryIntoFuture: cfg.MaxQueryIntoFuture,
 		}
 
 		dqr, err := distributor.Querier(ctx, mint, maxt)
@@ -263,8 +266,9 @@ type querier struct {
 	ctx         context.Context
 	mint, maxt  int64
 
-	tombstonesLoader *purger.TombstonesLoader
-	limits           *validation.Overrides
+	tombstonesLoader   *purger.TombstonesLoader
+	limits             *validation.Overrides
+	maxQueryIntoFuture time.Duration
 }
 
 // Select implements storage.Querier interface.
@@ -284,6 +288,8 @@ func (q querier) Select(_ bool, sp *storage.SelectHints, matchers ...*labels.Mat
 	// Also, in the recent versions of Prometheus, we pass in the hint but with Func set to "series".
 	// See: https://github.com/prometheus/prometheus/pull/8050
 	if sp == nil || sp.Func == "series" {
+		// In this case, the query time range has already been validated when the querier has been
+		// created.
 		return q.metadataQuerier.Select(true, sp, matchers...)
 	}
 
@@ -292,13 +298,30 @@ func (q querier) Select(_ bool, sp *storage.SelectHints, matchers ...*labels.Mat
 		return storage.ErrSeriesSet(err)
 	}
 
-	// Validate query time range.
-	startTime := model.Time(sp.Start)
-	endTime := model.Time(sp.End)
+	// Validate query time range. Even if the time range has already been validated when we created
+	// the querier, we need to check it again here because the time range specified in hints may be
+	// different.
+	startMs, endMs, err := validateQueryTimeRange(ctx, userID, sp.Start, sp.End, q.limits, q.maxQueryIntoFuture)
+	if err == errEmptyTimeRange {
+		return storage.NoopSeriesSet()
+	} else if err != nil {
+		return storage.ErrSeriesSet(err)
+	}
+
+	// The time range may have been manipulated during the validation,
+	// so we make sure changes are reflected back to hints.
+	sp.Start = startMs
+	sp.End = endMs
+
+	startTime := model.Time(startMs)
+	endTime := model.Time(endMs)
+
+	// Validate query time range. This validation should be done only for instant / range queries and
+	// NOT for metadata queries (series, labels) because the query-frontend doesn't support splitting
+	// of such queries.
 	if maxQueryLength := q.limits.MaxQueryLength(userID); maxQueryLength > 0 && endTime.Sub(startTime) > maxQueryLength {
 		limitErr := validation.LimitError(fmt.Sprintf(validation.ErrQueryTooLong, endTime.Sub(startTime), maxQueryLength))
 		return storage.ErrSeriesSet(limitErr)
-
 	}
 
 	tombstones, err := q.tombstonesLoader.GetPendingTombstonesForInterval(userID, startTime, endTime)
@@ -474,4 +497,42 @@ func UseBeforeTimestampQueryable(queryable storage.Queryable, ts time.Time) Quer
 		Queryable: queryable,
 		ts:        t,
 	}
+}
+
+func validateQueryTimeRange(ctx context.Context, userID string, startMs, endMs int64, limits *validation.Overrides, maxQueryIntoFuture time.Duration) (int64, int64, error) {
+	now := model.Now()
+	startTime := model.Time(startMs)
+	endTime := model.Time(endMs)
+
+	// Clamp time range based on max query into future.
+	if maxQueryIntoFuture > 0 && endTime.After(now.Add(maxQueryIntoFuture)) {
+		origEndTime := endTime
+		endTime = now.Add(maxQueryIntoFuture)
+
+		// Make sure to log it in traces to ease debugging.
+		level.Debug(spanlogger.FromContext(ctx)).Log(
+			"msg", "the end time of the query has been manipulated because of the 'max query into future' setting",
+			"original", origEndTime, "updated", endTime)
+
+		if endTime.Before(startTime) {
+			return 0, 0, errEmptyTimeRange
+		}
+	}
+
+	// Clamp the time range based on the max query lookback.
+	if maxQueryLookback := limits.MaxQueryLookback(userID); maxQueryLookback > 0 && startTime.Before(now.Add(-maxQueryLookback)) {
+		origStartTime := startTime
+		startTime = now.Add(-maxQueryLookback)
+
+		// Make sure to log it in traces to ease debugging.
+		level.Debug(spanlogger.FromContext(ctx)).Log(
+			"msg", "the start time of the query has been manipulated because of the 'max query lookback' setting",
+			"original", origStartTime, "updated", startTime)
+
+		if endTime.Before(startTime) {
+			return 0, 0, errEmptyTimeRange
+		}
+	}
+
+	return int64(startTime), int64(endTime), nil
 }